Finished implementing split_data Fixed path for tests
